feat: implement runtime configuration loading and refactor main function
